requirements.txt
boto3==1.4.4
botocore==1.5.60
Django==1.11.2
djangorestframework==3.6.3
docutils==0.13.1
jmespath==0.9.3
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
